feat: Add new rows in route
