+++ src/usecases/calledUseCases/index.ts
@@ -1,4 +1,4 @@
+import { CalledUseCase } from "./calledUsecases";
 import { prismaInstance } from "../../lib/prisma";
-import { CalledUseCase } from "./called.usecases";
 
 export const calledUseCase = new CalledUseCase(prismaInstance)
